rm repeat file
