+++ src/Makefile
@@ -1,10 +1,10 @@
 #$@--目标文件，$^--所有的依赖文件，$<--第一个依赖文件。
 target=calculator
-objects=main.o para.o character.o weapon.o artifact.o enemy.o calculate.o
-#flags=-fmessage-length=0 
+objects=origin.o artifact.o character.o enemy.o environment.o weapon.o  logger.o calculate.o main.o 
+flags=-std=c++11 -std=gnu++11
 #flags=''
 ${target}:${objects}
-	g++ ${flags} -o ${@} ${objects}
+	g++ ${flags} -o ${@} ${objects} 
 
 .PHONY : clean 
 clean:
