--- src/Makefile copy
@@ -1,11 +1,0 @@
-#$@--目标文件，$^--所有的依赖文件，$<--第一个依赖文件。
-target=calculator
-objects=origin.o artifact.o character.o enemy.o environment.o weapon.o  logger.o calculate.o main.o 
-flags=-std=c++11 -std=gnu++11
-#flags=''
-${target}:${objects}
-	g++ ${flags} -o ${@} ${objects} 
-
-.PHONY : clean 
-clean:
-	-rm ${target} ${objects}
