--- src/artifact copy.cpp
@@ -1,11 +1,0 @@
-#include "artifact.h"
-using namespace genShinImpact;
-using namespace std;
-artifact::artifact()
-{
-    memset(&base,'0',sizeof(tAllAttr));
-}
-artifact::~artifact()
-{    memset(&base,'0',sizeof(tAllAttr));
-
-}
