--- src/character copy.cpp
@@ -1,9 +1,0 @@
-#include "character.h"
-using namespace genShinImpact;
-using namespace std;
-character::character()
-{
-}
-character::~character()
-{
-}
