--- src/enemy copy.h
@@ -1,19 +1,0 @@
-#include "origin.h"
-#ifndef ENEMY_H
-#define ENEMY_H
-
-namespace genShinImpact
-{
-    class enemy
-    {
-    public:
-        tAllAttr base;
-        enemy();
-        ~enemy();
-
-    protected:
-    private:
-    };
-}
-
-#endif
