+++ src/enemy.cpp
@@ -1,19 +1,15 @@
 #include "enemy.h"
-#include "para.h"
 using namespace genShinImpact;
 using namespace std;
 enemy::enemy()
 {
-    memset(&mEnemy, '\0', sizeof(tEnemy));
-    mEnemy.level = 90;
-    INIT_STRUCT(mEnemy.res, tRes, float, 0.1);
-    mEnemy.res.hit=0;
-    mEnemy.res.chargeHit=0;
-    mEnemy.res.plunge=0;
-    mEnemy.res.skill=0;
-    mEnemy.res.burst=0;
-}
 
-enemy::enemy(tEnemy temp)
+    INIT_STRUCT(base.res, tRes, float, 0.1);
+    base.info.level = 90;
+    base.info.name = "Default Enemy";
+    base.info.rarity = RARITY_UNSURE;
+}
+enemy::~enemy()
 {
+    memset(&base, '0', sizeof(tAllAttr));
 }
