--- src/main copy.cpp
@@ -1,41 +1,0 @@
-//============================================================================
-// Name        : test.cpp
-// Author      :
-// Version     :
-// Copyright   : Your copyright notice
-// Description : Hello World in C++, Ansi-style
-//============================================================================
-#include "origin.h"
-
-#include "weapon.h"
-#include "artifact.h"
-#include "character.h"
-#include "enemy.h"
-#include "calculate.h"
-
-#include <iostream>
-#include <bitset> //输出二进制的头文件
-// cout<<hex/oct/dec/bin<<num<<endl;
-// cout<<bitset<sizeof(a)*8>(a)<<endl;
-#include <cstring>
-using namespace std;
-using namespace genShinImpact;
-//    A a(1);  //栈中分配
-//    A b = A(1);  //栈中分配
-//    A* c = new A(1);  //堆中分配
-
-int main()
-{
-
-	// character *testCharacter = new character();
-
-	// enemy *testEnemy = new enemy();
-	// weapon *testWeapon = new weapon();
-	// artifact *testArtifact = new artifact(ART_TEXT_MAIN_ATK, ART_TEXT_MAIN_ELEMENT_BONUS_CRYO, ART_TEXT_MAIN_CRIT_RATE, STAR_5); // ART_TEXT_MAIN_ELEMENT_BONUS
-	
-	// testCharacter->loadWeapon(&(testWeapon->mWeapon));
-	// testCharacter->loadTeamEffect(TEAM_EFFECT_DOUBLE_ELECTRO);
-	// testCharacter->mCharacter.attribute.elementalMastery=336;
-
-}													
-
