--- src/para.cpp
@@ -1,11 +1,0 @@
-#include "para.h"
-
-
-
-
-static void calFusionTable(int level, float *table)
-{
-}
-static void calArtifactTable(int type, float *table)
-{
-}
